Modified code to allow for end of 2023 season.
Now that the 2023 season is over, the code that ID's the next circuit is no longer working, as there are no rounds remainining in the season.
As a temporary measure, some of the code that is used to ID the next circuit has been commented out, and nextCircuitId is manually set to "bahrain" - the first
circuit of 2024.
Once the API updates the current season from "2023" to "2024", I will enable the commented code and remove the manual setting. Based on API documentation,
this is likely to be between 02/29/24 – 03/02/24