Adjusted style+timing of table loading animations
Edited the size and styling of the loading animations for table cells.
Adjusted placement of boolean state setters responsible for triggering the ending of the loading animations and rendering of the data. Confirmed that they are working as intended, as removing the state setters will render the loading animations indefinitely.
Note: Data renders almost instantly on my device, so animations are not appearing.