Continued work on getDriverAverages()
getDriverAverages() is now reporting average race finishes for every season of a driver's career, however it is merging the results together into a single array and deriving an average from that array.
At next update, will create a careerResults object for each driver with a prop that matches his lastName or driverId, then loop through each driver in the getDriverAverages function.
Added props for raceFinishes and meanRaceFinish to careerResults object.
In App.jsx, placed getDriverAverages() within a useEffect with empty dep array to ensure it only calls once.
Added a second driver's json data to data.json and placed each data set within an array to allow for looping within the getDriverAverages function.