Completed algorithm for average race finish
Reworked getDriverAverages() so it now takes in driverId (instead of lastName) as a variable for matching race results with drivers. This made it easier to access each driver's career results in data.json.
Last names are no longer be used for matching data with driver props, as it opens the possibility of future drivers sharing the same last name and causing issues with data.
Added props for driverId, allRaceResults, averageRaceFinishes, and careerData to each driver object. driverID is extracted from a simplified fetchCurrentStandings() function that only returns driver info. From there, a new useEffect in App.jsx extracts full names and driverIDs and sets them to state.
Moved getDriverAverages() out of utilities and into App.jsx, as it needs to reference and update several variables and props within App.jsx.